refactor: load Mirador from CDN
To avoid build-time dependency on all of Mirador's dependencies.
