+++ packages/mirador/package.json
@@ -3,18 +3,12 @@
   "version": "1.119.0",
   "description": "Mirador for Europeana website IIIF presentation",
   "license": "EUPL-1.2",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.js"
+  },
   "main": "src/index.js",
   "scripts": {
     "build": "npx webpack --config webpack.config.cjs"
-  },
-  "dependencies": {
-    "mirador": "^3.3.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
-  },
-  "devDependencies": {
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack": "4.46.0",
-    "webpack-cli": "^3.3.12"
   }
 }
