+++ packages/mirador/src/index.js
@@ -1,8 +1,26 @@
-import Mirador from 'mirador/dist/es/src/index.js';
 import theme from './theme.js';
 
-export default Mirador;
+export default {
+  data() {
+    return {
+      isMiradorLoaded: process.client ? !!window.Mirador : false,
+      miradorTheme: theme,
+      miradorViewer: null
+    };
+  },
 
-export {
-  theme
+  methods: {
+    async loadMirador() {
+      const script = document.createElement('script');
+      script.src = 'https://cdn.jsdelivr.net/npm/mirador@3.3.0/dist/mirador.min.js';
+      script.onload = this.initMirador;
+      document.head.appendChild(script);
+
+      this.isMiradorLoaded = true;
+    },
+
+    async initMirador() {
+      this.miradorViewer = window.Mirador.viewer(this.miradorViewerOptions, this.miradorViewerPlugins);
+    }
+  }
 };
